pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nifty-trader"
version = "0.1.0"
description = "NIFTY 50 Options Directional Buying System"
requires-python = ">=3.11"
dependencies = [
    "dhanhq>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "numpy>=1.26",
    "pandas>=2.1",
    "rich>=13.0",
    "httpx>=0.27",
    "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[project.scripts]
nifty-trader = "nifty_trader.main:main"

[tool.setuptools.packages.find]
where = ["src"]
